+++ kernel_tuner/strategies/firefly_algorithm.py
@@ -30,7 +30,6 @@ def tune(runner, kernel_options, device_options, tuning_options):
     """
 
     results = []
-    cache = {}
 
     #scale variables in x because PSO works with velocities to visit different configurations
     tuning_options["scaling"] = True
@@ -38,15 +37,15 @@ def tune(runner, kernel_options, device_options, tuning_options):
     #using this instead of get_bounds because scaling is used
     bounds, _, _ = get_bounds_x0_eps(tuning_options)
 
-    args = (kernel_options, tuning_options, runner, results, cache)
+    args = (kernel_options, tuning_options, runner, results)
 
-    num_particles = 20
-    maxiter = 100
+    num_particles = tuning_options.strategy_options.get("popsize", 20)
+    maxiter = tuning_options.strategy_options.get("maxiter", 100)
 
     #parameters needed by the Firefly Algorithm
-    B0 = 1.0
-    gamma = 1.0
-    alpha = 0.20
+    B0 = tuning_options.strategy_options.get("B0", 1.0)
+    gamma = tuning_options.strategy_options.get("gamma", 1.0)
+    alpha = tuning_options.strategy_options.get("alpha", 0.2)
 
     best_time_global = 1e20
     best_position_global = []